requirements/base.txt
arrow==0.10.0
factory-boy==2.8.1
Faker==0.7.7
invoke==0.14.0
Pillow==3.4.2
python-dateutil==2.6.0
requests==2.12.4
six==1.10.0
